Rename Menu.tsx -> MenuComponent.tsx and wire up to root component
