+++ browser/src/Services/Menu/index.ts
@@ -1,1 +1,2 @@
 export * from "./Menu"
+export * from "./MenuComponent"
